requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
filelock==3.8.0
fonttools==4.36.0
frozenlist==1.3.1
fsspec==2022.7.1
google-auth==2.10.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h5py==3.7.0
huggingface-hub==0.9.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==7.34.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.11.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.21.6
oauthlib==3.2.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyfaidx==0.7.1
Pygments==2.13.0
pyparsing==3.0.9
pytabix==0.1
python-dateutil==2.8.2
pytorch-lightning==1.7.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
selene-sdk==0.5.0
six==1.16.0
statsmodels==0.13.2
tabulate==0.8.10
tenacity==8.0.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.1
torchmetrics==0.9.3
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.2
typing-extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
